fix: bug in ownerInfo controller and check iran phone

+++ src/controllers/ownerInfo.ts
@@ -4,7 +4,11 @@ import httpErrors from "http-errors";
 import { deleteFile, uploadFile } from "../utils/s3";
 import httpStatus from "http-status";
 
-export const getOne = async (req: Request, res: Response, next: NextFunction) => {
+export const getOne = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
   try {
     const ownerInfo = await OwnerInfoModel.findOne();
 
@@ -14,7 +18,11 @@ export const getOne = async (req: Request, res: Response, next: NextFunction) =>
   }
 };
 
-export const create = async (req: Request, res: Response, next: NextFunction) => {
+export const create = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
   try {
     const { body } = req;
 
@@ -56,7 +64,11 @@ export const create = async (req: Request, res: Response, next: NextFunction) =>
   }
 };
 
-export const update = async (req: Request, res: Response, next: NextFunction) => {
+export const update = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
   try {
     const { body } = req;
 
@@ -76,7 +88,8 @@ export const update = async (req: Request, res: Response, next: NextFunction) =>
       const path = `/avatar/${Date.now()}--${avatar[0].originalname}`;
       const uploadedAvatar = await uploadFile(avatar[0].buffer, path);
       avatarLocation = uploadedAvatar.Location;
-      await deleteFile(ownerInfo.dataValues.avatarURL);
+      if (ownerInfo.dataValues.avatarURL)
+        await deleteFile(ownerInfo.dataValues.avatarURL);
     }
 
     let coverLocation: null | string = null;
@@ -84,7 +97,8 @@ export const update = async (req: Request, res: Response, next: NextFunction) =>
       const path = `/cover/${Date.now()}--${cover[0].originalname}`;
       const uploadedCover = await uploadFile(cover[0].buffer, path);
       coverLocation = uploadedCover.Location;
-      await deleteFile(ownerInfo.dataValues.mainCover);
+      if (ownerInfo.dataValues.mainCover)
+        await deleteFile(ownerInfo.dataValues.mainCover);
     }
 
     await OwnerInfoModel.update(
